Migrating user_client_ca_test to ginkgo
